Add jobs search
Add get_jobs functions, which makes requests to The Muse API based on
filter parameters, and returns the results.

Add POST method to the jobs route, taking parameters based on what the
client selects from dropdown menus.

Add a default 'All' option to the dropdown menus.

If the client submits filters, feed the selected data back into the
template, allowing for multiple fields to be selected.